@@ -10,100 +10,98 @@
 		"preview": "vite preview"
 	},
 	"dependencies": {
-		"@babel/core": "^7.25.2",
+		"@babel/core": "^7.26.0",
 		"@fingerprintjs/fingerprintjs-pro-react": "^2.6.3",
-		"@hookform/resolvers": "^3.9.0",
-		"@radix-ui/react-accordion": "^1.2.0",
-		"@radix-ui/react-alert-dialog": "^1.1.1",
+		"@hookform/resolvers": "^3.9.1",
+		"@radix-ui/react-accordion": "^1.2.1",
+		"@radix-ui/react-alert-dialog": "^1.1.2",
 		"@radix-ui/react-aspect-ratio": "^1.1.0",
-		"@radix-ui/react-avatar": "^1.1.0",
-		"@radix-ui/react-checkbox": "^1.1.1",
-		"@radix-ui/react-collapsible": "^1.1.0",
-		"@radix-ui/react-context-menu": "^2.2.1",
-		"@radix-ui/react-dialog": "^1.1.1",
-		"@radix-ui/react-dropdown-menu": "^2.1.1",
-		"@radix-ui/react-hover-card": "^1.1.1",
-		"@radix-ui/react-icons": "^1.3.0",
+		"@radix-ui/react-avatar": "^1.1.1",
+		"@radix-ui/react-checkbox": "^1.1.2",
+		"@radix-ui/react-collapsible": "^1.1.1",
+		"@radix-ui/react-context-menu": "^2.2.2",
+		"@radix-ui/react-dialog": "^1.1.2",
+		"@radix-ui/react-dropdown-menu": "^2.1.2",
+		"@radix-ui/react-hover-card": "^1.1.2",
+		"@radix-ui/react-icons": "^1.3.2",
 		"@radix-ui/react-label": "^2.1.0",
-		"@radix-ui/react-menubar": "^1.1.1",
-		"@radix-ui/react-navigation-menu": "^1.2.0",
-		"@radix-ui/react-popover": "^1.1.1",
+		"@radix-ui/react-menubar": "^1.1.2",
+		"@radix-ui/react-navigation-menu": "^1.2.1",
+		"@radix-ui/react-popover": "^1.1.2",
 		"@radix-ui/react-progress": "^1.1.0",
-		"@radix-ui/react-radio-group": "^1.2.0",
-		"@radix-ui/react-scroll-area": "^1.1.0",
-		"@radix-ui/react-select": "^2.1.1",
+		"@radix-ui/react-radio-group": "^1.2.1",
+		"@radix-ui/react-scroll-area": "^1.2.1",
+		"@radix-ui/react-select": "^2.1.2",
 		"@radix-ui/react-separator": "^1.1.0",
-		"@radix-ui/react-slider": "^1.2.0",
+		"@radix-ui/react-slider": "^1.2.1",
 		"@radix-ui/react-slot": "^1.1.0",
-		"@radix-ui/react-switch": "^1.1.0",
-		"@radix-ui/react-tabs": "^1.1.0",
-		"@radix-ui/react-toast": "^1.2.1",
+		"@radix-ui/react-switch": "^1.1.1",
+		"@radix-ui/react-tabs": "^1.1.1",
+		"@radix-ui/react-toast": "^1.2.2",
 		"@radix-ui/react-toggle": "^1.1.0",
 		"@radix-ui/react-toggle-group": "^1.1.0",
-		"@radix-ui/react-tooltip": "^1.1.2",
-		"@tanstack/react-query": "^5.56.2",
-		"@tanstack/react-virtual": "^3.10.8",
+		"@radix-ui/react-tooltip": "^1.1.4",
+		"@tanstack/react-query": "^5.61.5",
+		"@tanstack/react-virtual": "^3.10.9",
+		"@vscode/codicons": "^0.0.36",
 		"@vscode/webview-ui-toolkit": "^1.4.0",
 		"@xterm/addon-fit": "^0.10.0",
 		"@xterm/addon-web-links": "^0.11.0",
 		"@xterm/xterm": "^5.5.0",
-		"class-variance-authority": "^0.7.0",
+		"class-variance-authority": "^0.7.1",
 		"clsx": "^2.1.1",
-		"cmdk": "^1.0.0",
+		"cmdk": "^1.0.4",
 		"color": "^4.2.3",
 		"date-fns": "^3.6.0",
-		"embla-carousel-react": "^8.2.1",
-		"framer-motion": "^11.5.4",
+		"embla-carousel-react": "^8.5.1",
+		"framer-motion": "^11.12.0",
 		"fuse.js": "^7.0.0",
-		"input-otp": "^1.2.4",
-		"jotai-devtools": "^0.10.1",
+		"input-otp": "^1.4.1",
 		"jsdom": "^25.0.1",
 		"lucide-react": "^0.439.0",
 		"next-themes": "^0.3.0",
 		"react": "^18.3.1",
 		"react-day-picker": "^8.10.1",
 		"react-dom": "^18.3.1",
-		"react-hook-form": "^7.53.0",
+		"react-hook-form": "^7.53.2",
 		"react-markdown": "^9.0.1",
-		"react-resizable-panels": "^2.1.2",
-		"react-syntax-highlighter": "^15.5.0",
-		"react-textarea-autosize": "^8.5.3",
+		"react-resizable-panels": "^2.1.7",
+		"react-syntax-highlighter": "^15.6.1",
+		"react-textarea-autosize": "^8.5.5",
 		"react-use": "^17.5.1",
 		"react-virtualized-auto-sizer": "^1.0.24",
-		"react-virtuoso": "^4.7.13",
+		"react-virtuoso": "^4.12.2",
 		"react-window": "^1.8.10",
 		"react-wrap-balancer": "^1.1.1",
-		"recharts": "^2.12.7",
+		"recharts": "^2.13.3",
 		"rehype-raw": "^7.0.0",
 		"remark-gfm": "^4.0.0",
-		"sonner": "^1.5.0",
+		"sonner": "^1.7.0",
 		"styled-components": "^6.1.13",
-		"tailwind-merge": "^2.5.2",
+		"tailwind-merge": "^2.5.5",
 		"tailwindcss-animate": "^1.0.7",
 		"use-resize-observer": "^9.1.0",
-		"vaul": "^0.9.2",
-		"web-vitals": "^4.2.3",
-		"xterm": "^5.3.0",
-		"xterm-addon-fit": "^0.8.0",
+		"vaul": "^0.9.9",
+		"web-vitals": "^4.2.4",
 		"zod": "^3.23.8"
 	},
 	"devDependencies": {
 		"@types/jsdom": "^21.1.7",
-		"@types/node": "^22.5.4",
-		"@types/react": "^18.3.3",
-		"@types/react-dom": "^18.3.0",
+		"@types/node": "^22.10.0",
+		"@types/react": "^18.3.12",
+		"@types/react-dom": "^18.3.1",
 		"@types/react-syntax-highlighter": "^15.5.13",
 		"@types/vscode-webview": "^1.57.5",
 		"@types/xmldom": "^0.1.34",
-		"@vitejs/plugin-react": "^4.3.1",
-		"@vitejs/plugin-react-swc": "^3.7.1",
+		"@vitejs/plugin-react": "^4.3.4",
+		"@vitejs/plugin-react-swc": "^3.7.2",
 		"autoprefixer": "^10.4.20",
-		"eslint": "^9.9.0",
-		"eslint-plugin-react-hooks": "^5.1.0-rc.0",
-		"eslint-plugin-react-refresh": "^0.4.9",
-		"postcss": "^8.4.45",
-		"tailwindcss": "^3.4.10",
-		"typescript": "^5.5.3",
-		"vite": "^5.4.1"
+		"eslint": "^9.15.0",
+		"eslint-plugin-react-hooks": "5.1.0-rc-fb9a90fa48-20240614",
+		"eslint-plugin-react-refresh": "^0.4.14",
+		"postcss": "^8.4.49",
+		"tailwindcss": "^3.4.15",
+		"typescript": "^5.7.2",
+		"vite": "^6.0.1"
 	}
 }
