@@ -1,11 +1,10 @@
 import { Anthropic } from "@anthropic-ai/sdk"
 import { ExtensionProvider } from "../../providers/claude-coder/ClaudeCoderProvider"
 import { isV1ClaudeMessage } from "../../shared/ExtensionMessage"
-import { ToolName } from "../../shared/Tool"
 import { ClaudeAskResponse } from "../../shared/WebviewMessage"
 import { ApiManager } from "./api-handler"
 import { ToolExecutor } from "./tools/tool-executor"
-import { KoduDevOptions, ToolResponse, UserContent } from "./types"
+import { KoduDevOptions, ToolName, ToolResponse, UserContent } from "./types"
 import { getCwd, formatImagesIntoBlocks, getPotentiallyRelevantDetails, formatFilesList } from "./utils"
 import { StateManager } from "./state-manager"
 import { findLastIndex } from "../../utils"
