@@ -27,6 +27,7 @@ export type GlobalState = {
 	useUdiff: boolean | undefined
 	alwaysAllowReadOnly: boolean | undefined
 	alwaysAllowWriteOnly: boolean | undefined
+	inlineEditOutputType?: "full" | "diff" | "none"
 	autoSummarize: boolean | undefined
 	taskHistory: HistoryItem[] | undefined
 	shouldShowKoduPromo: boolean | undefined
@@ -38,6 +39,7 @@ export type GlobalState = {
 	technicalBackground: "no-technical" | "technical" | "developer" | undefined
 	systemPromptVariants: SystemPromptVariant[] | undefined
 	activeSystemPromptVariantId: string | undefined
+	commandTimeout: number | undefined
 }
 
 export class GlobalStateManager {
@@ -66,7 +68,7 @@ export class GlobalStateManager {
 
 			// Import and update prompt files
 			const promptFiles = {
-				"m-11-1-2024": () => import("../../../agent/v1/prompts/m-11-1-2024.prompt"),
+				"m-20-1-2024": () => import("../../../agent/v1/prompts/m-11-20-2024.prompt"),
 				"c-11-1-2024": () => import("../../../agent/v1/prompts/c-11-1-2024.prompt"),
 			}
 
@@ -101,7 +103,7 @@ export class GlobalStateManager {
 			await this.updateGlobalState("systemPromptVariants", variants)
 			// now check if the active system prompt variant is set and is valid otherwise set m-11-1-2024 as the active system prompt variant
 			if (!activeId || !variants.find((v) => v.id === activeId)) {
-				await this.updateGlobalState("activeSystemPromptVariantId", "m-11-1-2024")
+				await this.updateGlobalState("activeSystemPromptVariantId", "m-11-20-2024")
 			}
 		} catch (error) {
 			console.error("Error initializing system prompts:", error)
