@@ -1,8 +1,8 @@
-import * as path from "node:path";
-import { runTests } from "@vscode/test-electron";
-import { fileURLToPath } from "node:url";
-
-const __dirname = path.dirname(fileURLToPath(import.meta.url));
+import * as path from "node:path"
+import { runTests } from "@vscode/test-electron"
+import { fileURLToPath } from "node:url"
+import * as vscode from "vscode"
+const __dirname = path.dirname(fileURLToPath(import.meta.url))
 
 async function main() {
 	try {
@@ -11,7 +11,8 @@ async function main() {
 
 		// Path to the test suite
 		const extensionTestsPath = path.resolve(__dirname, "./suite")
-
+		const extension = vscode.extensions.getExtension("kodu-ai.claude-dev-experimental")!
+		await extension.activate()
 		// Run the integration tests
 		await runTests({
 			extensionDevelopmentPath,
