@@ -1,11 +1,11 @@
+import { Input } from "@/components/ui/input"
 import { Label } from "@/components/ui/label"
 import { Textarea } from "@/components/ui/textarea"
-import { Input } from "@/components/ui/input"
-import React, { useEffect } from "react"
+import React from "react"
 import { useSettingsState } from "../../hooks/useSettingsState"
+import { Slider } from "../ui/slider"
 import { ExperimentalFeatureItem } from "./experimental-feature-item"
 import SystemPromptVariants from "./SystemPromptVariants"
-import { Slider } from "../ui/slider"
 
 const AdvancedTab: React.FC = () => {
 	const {
@@ -14,12 +14,27 @@ const AdvancedTab: React.FC = () => {
 		autoSkipWrite,
 		customInstructions,
 		terminalCompressionThreshold,
+		commandTimeout,
+		handleCommandTimeout,
 		handleTerminalCompressionThresholdChange,
 		handleSetReadOnly,
 		handleSetAutoCloseTerminal,
 		handleAutoSkipWriteChange,
 		handleCustomInstructionsChange,
 	} = useSettingsState()
+
+	const handleTextAreaChange = (e: React.ChangeEvent<HTMLTextAreaElement>) => {
+		const textarea = e.target
+		const cursorPosition = textarea.selectionStart
+
+		handleCustomInstructionsChange(e.target.value)
+
+		// Restore cursor position after state update
+		requestAnimationFrame(() => {
+			textarea.selectionStart = cursorPosition
+			textarea.selectionEnd = cursorPosition
+		})
+	}
 
 	return (
 		<div className="space-y-6">
@@ -58,7 +73,7 @@ const AdvancedTab: React.FC = () => {
 							id: "terminalCompressionThreshold",
 							label: "Enable Terminal Compression",
 							description:
-								"Compress terminal output to reduce token usage when the output exceeds the threshold",
+								"Compress terminal output to reduce token usage when the output exceeds the threshold at the end of context window",
 						}}
 						checked={terminalCompressionThreshold !== undefined}
 						onCheckedChange={(checked) =>
@@ -107,7 +122,44 @@ const AdvancedTab: React.FC = () => {
 					)}
 				</div>
 			</div>
-
+			<div className="space-y-4 mx-0">
+				<div className="pl-0 grid gap-4">
+					<div className="grid gap-2">
+						<Label htmlFor="range">Command Timeout</Label>
+						<div className="grid gap-4">
+							<div className="flex items-center gap-4">
+								<Input
+									id="command-timeout"
+									type="number"
+									value={commandTimeout ?? 30}
+									onChange={(e) => {
+										const value = parseInt(e.target.value)
+										if (!isNaN(value)) {
+											handleCommandTimeout(value)
+										}
+									}}
+									min={10}
+									max={600}
+									step={10}
+									className="w-24"
+								/>
+								<span className="text-sm text-muted-foreground">(10 - 600)</span>
+							</div>
+							<Slider
+								min={10}
+								max={600}
+								step={10}
+								value={[commandTimeout ?? 30]}
+								onValueChange={(value) => handleCommandTimeout(value[0])}
+								className="w-full"
+							/>
+						</div>
+						<p className="text-sm text-muted-foreground">
+							Set the maximum time in seconds that a command can run before being terminated
+						</p>
+					</div>
+				</div>
+			</div>
 			<div className="space-y-2">
 				<Label htmlFor="custom-instructions" className="text-xs font-medium">
 					Custom Instructions
@@ -116,9 +168,14 @@ const AdvancedTab: React.FC = () => {
 					id="custom-instructions"
 					placeholder="e.g. 'Run unit tests at the end', 'Use TypeScript with async/await'"
 					value={customInstructions}
-					onChange={(e) => handleCustomInstructionsChange(e.target.value)}
-					className="min-h-[60px] text-xs"
+					onChange={handleTextAreaChange}
+					className="min-h-[120px] text-xs resize-y"
+					style={{
+						fontFamily: "var(--vscode-editor-font-family)",
+					}}
+					spellCheck={false}
 				/>
+				<p className="text-xs text-muted-foreground mt-1">These instructions will be included in every task</p>
 			</div>
 
 			<SystemPromptVariants />
