@@ -1,8 +1,8 @@
+import "@vscode/codicons/dist/codicon.css"
 import React from "react"
 import ReactDOM from "react-dom/client"
-import "./index.css"
 import App from "./App"
-import "../../node_modules/@vscode/codicons/dist/codicon.css"
+import "./index.css"
 
 const root = ReactDOM.createRoot(document.getElementById("root") as HTMLElement)
 root.render(
